refactor(outline-item): outline 관련 context와 hook을 outline 내부로 이동
